# specs/001-puphax-rest-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PUPHAX REST API Service
  description: |
    REST API service that provides access to the Hungarian NEAK PUPHAX drug database.
    
    This service converts SOAP responses from the PUPHAX public drug database into 
    structured JSON format, providing healthcare applications with easy access to 
    official drug information including active ingredients, manufacturers, ATC codes,
    and regulatory details.
    
    **Features:**
    - Drug search by name, manufacturer, or ATC code
    - Paginated results with configurable page sizes
    - Response caching for improved performance
    - Comprehensive error handling
    - Production-ready monitoring and health checks
    
    **Data Source:** NEAK PUPHAX Public Drug Database  
    **Update Frequency:** Real-time via SOAP service
  version: 1.0.0
  contact:
    name: PUPHAX REST API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /api/v1/drugs/search:
    get:
      tags:
        - Drug Search
      summary: Search for drugs in the PUPHAX database
      description: |
        Search for pharmaceutical products using various criteria including drug name,
        manufacturer, or ATC code. Results are paginated and can be sorted by different fields.
        
        **Search Strategy:**
        - Primary search by drug name or active ingredient
        - Optional filtering by manufacturer and ATC code
        - Case-insensitive search with Hungarian character support
        - Partial matching supported for drug names
        
        **Performance:**
        - Results are cached for 5-15 minutes
        - Typical response time: <2 seconds
        - Supports up to 100 concurrent requests
      operationId: searchDrugs
      parameters:
        - name: term
          in: query
          required: true
          description: Primary search term (drug name or active ingredient)
          schema:
            type: string
            minLength: 2
            maxLength: 100
          example: "aspirin"
        - name: manufacturer
          in: query
          required: false
          description: Filter by manufacturer name
          schema:
            type: string
            maxLength: 100
          example: "Bayer"
        - name: atcCode
          in: query
          required: false
          description: Filter by ATC code (Anatomical Therapeutic Chemical)
          schema:
            type: string
            pattern: "^[A-Z][0-9]{2}[A-Z]{2}[0-9]{2}$"
          example: "N02BA01"
        - name: page
          in: query
          required: false
          description: Page number for pagination (0-based)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          required: false
          description: Number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: sortBy
          in: query
          required: false
          description: Field to sort results by
          schema:
            type: string
            enum: [name, manufacturer, atcCode]
            default: name
          example: "name"
        - name: sortDirection
          in: query
          required: false
          description: Sort direction
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
          example: "ASC"
      responses:
        '200':
          description: Successful search operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrugSearchResponse'
              examples:
                successful_search:
                  summary: Successful search with results
                  value:
                    drugs:
                      - id: "HU001234"
                        name: "Aspirin 100mg"
                        manufacturer: "Bayer Hungary Kft."
                        atcCode: "N02BA01"
                        activeIngredients: ["Acetylsalicylic acid"]
                        prescriptionRequired: false
                        reimbursable: true
                        status: "ACTIVE"
                      - id: "HU005678"
                        name: "Aspirin Protect 100mg"
                        manufacturer: "Bayer Hungary Kft."
                        atcCode: "N02BA01"
                        activeIngredients: ["Acetylsalicylic acid"]
                        prescriptionRequired: false
                        reimbursable: true
                        status: "ACTIVE"
                    pagination:
                      currentPage: 0
                      pageSize: 20
                      totalPages: 1
                      totalElements: 2
                      hasNext: false
                      hasPrevious: false
                    searchInfo:
                      searchTerm: "aspirin"
                      filters: {}
                      executionTime: 245
                      cacheHit: false
                      timestamp: "2025-10-20T10:30:00Z"
                empty_results:
                  summary: Search with no results
                  value:
                    drugs: []
                    pagination:
                      currentPage: 0
                      pageSize: 20
                      totalPages: 0
                      totalElements: 0
                      hasNext: false
                      hasPrevious: false
                    searchInfo:
                      searchTerm: "nonexistentdrug"
                      filters: {}
                      executionTime: 89
                      cacheHit: true
                      timestamp: "2025-10-20T10:30:00Z"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                missing_term:
                  summary: Missing required search term
                  value:
                    timestamp: "2025-10-20T10:30:00Z"
                    status: 400
                    error: "Validation Failed"
                    message: "Request validation failed"
                    path: "/api/v1/drugs/search"
                    correlationId: "req-123456"
                    fieldErrors:
                      - field: "term"
                        rejectedValue: null
                        message: "Search term is required"
                invalid_atc:
                  summary: Invalid ATC code format
                  value:
                    timestamp: "2025-10-20T10:30:00Z"
                    status: 400
                    error: "Validation Failed"
                    message: "Request validation failed"
                    path: "/api/v1/drugs/search"
                    correlationId: "req-123457"
                    fieldErrors:
                      - field: "atcCode"
                        rejectedValue: "INVALID"
                        message: "ATC code must follow format: letter-2digits-2letters-2digits"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                soap_service_error:
                  summary: PUPHAX SOAP service unavailable
                  value:
                    timestamp: "2025-10-20T10:30:00Z"
                    status: 500
                    error: "Service Unavailable"
                    message: "PUPHAX service is temporarily unavailable. Please try again later."
                    path: "/api/v1/drugs/search"
                    correlationId: "req-123458"
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                circuit_breaker_open:
                  summary: Circuit breaker activated
                  value:
                    timestamp: "2025-10-20T10:30:00Z"
                    status: 503
                    error: "Circuit Breaker Open"
                    message: "Service is temporarily unavailable due to repeated failures. Please try again in 30 seconds."
                    path: "/api/v1/drugs/search"
                    correlationId: "req-123459"

  /api/v1/drugs/{drugId}:
    get:
      tags:
        - Drug Details
      summary: Get detailed information about a specific drug
      description: |
        Retrieve comprehensive information about a specific drug using its unique identifier.
        This endpoint provides complete drug details including all active ingredients,
        dosage forms, regulatory information, and current status.
      operationId: getDrugById
      parameters:
        - name: drugId
          in: path
          required: true
          description: Unique drug identifier from PUPHAX database
          schema:
            type: string
          example: "HU001234"
      responses:
        '200':
          description: Drug found and details returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrugRecord'
        '404':
          description: Drug not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /actuator/health:
    get:
      tags:
        - Health Checks
      summary: Service health status
      description: |
        Check the health status of the service including connectivity to the PUPHAX SOAP service.
        Returns detailed health information including cache status, SOAP service connectivity,
        and overall service status.
      operationId: getHealthStatus
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [UP, DOWN]
                  components:
                    type: object
                    properties:
                      puphaxSoap:
                        type: object
                        properties:
                          status:
                            type: string
                            enum: [UP, DOWN]
                          details:
                            type: object
                            properties:
                              lastCheck:
                                type: string
                                format: date-time
                              responseTime:
                                type: integer
                      cache:
                        type: object
                        properties:
                          status:
                            type: string
                            enum: [UP, DOWN]
                          details:
                            type: object
                            properties:
                              hitRatio:
                                type: number
                              totalEntries:
                                type: integer

components:
  schemas:
    DrugSearchResponse:
      type: object
      required:
        - drugs
        - pagination
        - searchInfo
      properties:
        drugs:
          type: array
          items:
            $ref: '#/components/schemas/DrugSummary'
          description: List of drugs matching search criteria
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        searchInfo:
          $ref: '#/components/schemas/SearchInfo'

    DrugSummary:
      type: object
      required:
        - id
        - name
        - manufacturer
        - status
      properties:
        id:
          type: string
          description: Unique drug identifier
          example: "HU001234"
        name:
          type: string
          description: Commercial drug name
          example: "Aspirin 100mg"
        manufacturer:
          type: string
          description: Pharmaceutical company name
          example: "Bayer Hungary Kft."
        atcCode:
          type: string
          description: ATC classification code
          example: "N02BA01"
        activeIngredients:
          type: array
          items:
            type: string
          description: List of active ingredient names
          example: ["Acetylsalicylic acid"]
        prescriptionRequired:
          type: boolean
          description: Whether prescription is required
          example: false
        reimbursable:
          type: boolean
          description: Whether covered by insurance
          example: true
        status:
          $ref: '#/components/schemas/DrugStatus'

    DrugRecord:
      type: object
      required:
        - id
        - name
        - activeIngredients
        - manufacturer
        - status
      properties:
        id:
          type: string
          description: Unique drug identifier
          example: "HU001234"
        name:
          type: string
          description: Commercial drug name
          example: "Aspirin 100mg enteric-coated tablets"
        activeIngredients:
          type: array
          items:
            $ref: '#/components/schemas/ActiveIngredient'
          description: Active pharmaceutical ingredients
        manufacturer:
          type: string
          description: Pharmaceutical company name
          example: "Bayer Hungary Kft."
        atcCode:
          type: string
          description: ATC classification code
          example: "N02BA01"
        therapeuticGroup:
          type: string
          description: Therapeutic category
          example: "Analgesics and antipyretics"
        dosageForms:
          type: array
          items:
            $ref: '#/components/schemas/DosageForm'
          description: Available dosage forms
        strength:
          type: string
          description: Drug strength/concentration
          example: "100mg"
        packagingSize:
          type: string
          description: Package size information
          example: "30 tablets"
        prescriptionRequired:
          type: boolean
          description: Whether prescription is required
          example: false
        reimbursable:
          type: boolean
          description: Whether covered by insurance
          example: true
        registrationNumber:
          type: string
          description: Official registration number
          example: "OGYI-T-20123/01"
        registrationDate:
          type: string
          format: date
          description: Date of registration
          example: "2015-03-15"
        expiryDate:
          type: string
          format: date
          description: Registration expiry date
          example: "2025-03-15"
        status:
          $ref: '#/components/schemas/DrugStatus'
        additionalInfo:
          type: string
          description: Additional regulatory information
          example: "Suitable for long-term cardiovascular protection"

    ActiveIngredient:
      type: object
      required:
        - name
        - concentration
        - role
      properties:
        name:
          type: string
          description: Chemical name of active ingredient
          example: "Acetylsalicylic acid"
        concentration:
          type: string
          description: Concentration amount and unit
          example: "100mg"
        role:
          $ref: '#/components/schemas/IngredientRole'

    DosageForm:
      type: object
      required:
        - form
        - route
      properties:
        form:
          type: string
          description: Dosage form type
          example: "enteric-coated tablet"
        route:
          type: string
          description: Route of administration
          example: "oral"
        description:
          type: string
          description: Additional form description
          example: "Gastro-resistant coating"

    PaginationInfo:
      type: object
      required:
        - currentPage
        - pageSize
        - totalPages
        - totalElements
        - hasNext
        - hasPrevious
      properties:
        currentPage:
          type: integer
          description: Current page number (0-based)
          example: 0
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
        totalElements:
          type: integer
          format: int64
          description: Total number of matching records
          example: 45
        hasNext:
          type: boolean
          description: Whether there are more pages
          example: true
        hasPrevious:
          type: boolean
          description: Whether there are previous pages
          example: false

    SearchInfo:
      type: object
      required:
        - searchTerm
        - filters
        - executionTime
        - cacheHit
        - timestamp
      properties:
        searchTerm:
          type: string
          description: Original search term
          example: "aspirin"
        filters:
          type: object
          additionalProperties:
            type: string
          description: Applied filters
          example:
            manufacturer: "Bayer"
        executionTime:
          type: integer
          format: int64
          description: Search execution time in milliseconds
          example: 245
        cacheHit:
          type: boolean
          description: Whether result was served from cache
          example: false
        timestamp:
          type: string
          format: date-time
          description: When search was performed
          example: "2025-10-20T10:30:00Z"

    ApiErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
        - correlationId
      properties:
        timestamp:
          type: string
          format: date-time
          description: When error occurred
          example: "2025-10-20T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 500
        error:
          type: string
          description: Error type/category
          example: "Internal Server Error"
        message:
          type: string
          description: Human-readable error message
          example: "PUPHAX service is temporarily unavailable"
        path:
          type: string
          description: Request path that caused error
          example: "/api/v1/drugs/search"
        correlationId:
          type: string
          description: Unique identifier for request tracing
          example: "req-123456"

    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
              description: Specific field validation errors

    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Name of field that failed validation
          example: "term"
        rejectedValue:
          description: Value that was rejected
          example: null
        message:
          type: string
          description: Validation error message
          example: "Search term is required"

    DrugStatus:
      type: string
      enum:
        - ACTIVE
        - SUSPENDED
        - WITHDRAWN
      description: Current drug status
      example: "ACTIVE"

    IngredientRole:
      type: string
      enum:
        - ACTIVE
        - EXCIPIENT
      description: Role of ingredient in formulation
      example: "ACTIVE"

  securitySchemes: {}

tags:
  - name: Drug Search
    description: Operations for searching drugs in the PUPHAX database
  - name: Drug Details
    description: Operations for retrieving detailed drug information
  - name: Health Checks
    description: Service health and monitoring endpoints